my-c1-app/my-c1-project/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.11.0
openai==1.66.3
python-dotenv==1.0.1
crayonai_stream==0.6.2
thesys_genui_sdk==0.1.2
httpx==0.28.1
pymongo==4.10.1
